requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# Database
supabase==2.11.0

# ML / Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.11.0

# Image processing
rasterio>=1.3.9
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0
